Throw an IOException for requests to the mail/send endpoint if a template id in the request body doesn't exist in the user's account.
This is enabled by setting a SENDGRID_CHECK_TEMPLATES environment variable to "true".

A new package-private SendGrid(String apiKey, Client client, Map<String, String> environment) constructor was added to facilitate testing.

The other SendGrid constructors were also modified so that initialisation of the instance variables is only done in one place, i.e. in the new constructor.

A new SendGridTemplateIdTest class was added to test the new functionality.  It doesn't require prism to be running as the client is mocked.